+++ tests/tests.rs
@@ -3963,16 +3963,41 @@ mod borsh_tests {
         // Now try various arbitrary ints
         test_roundtrip(u1::new(0b0), &[0]);
         test_roundtrip(u1::new(0b1), &[1]);
+        test_roundtrip(i1::new(0), &[0]);
+        test_roundtrip(i1::new(-1), &[1]);
+
         test_roundtrip(u6::new(0b101101), &[0b101101]);
+        test_roundtrip(i6::from_bits(0b101101), &[0b101101]);
+
         test_roundtrip(u14::new(0b110101_11001101), &[0b11001101, 0b110101]);
+        test_roundtrip(i14::from_bits(0b110101_11001101), &[0b11001101, 0b110101]);
+
         test_roundtrip(
             u72::new(0x36_01234567_89ABCDEF),
             &[0xEF, 0xCD, 0xAB, 0x89, 0x67, 0x45, 0x23, 0x01, 0x36],
         );
+        test_roundtrip(
+            i72::from_bits(0x36_01234567_89ABCDEF),
+            &[0xEF, 0xCD, 0xAB, 0x89, 0x67, 0x45, 0x23, 0x01, 0x36],
+        );
+
         test_roundtrip(u24::MAX, &u24::MAX.to_le_bytes());
+        test_roundtrip(i24::MAX, &i24::MAX.to_le_bytes());
+
         test_roundtrip(u24::MIN, &u24::MIN.to_le_bytes());
+        test_roundtrip(i24::MIN, &i24::MIN.to_le_bytes());
+
         test_roundtrip(u24::new(0xAB_CD_EF), &u24::new(0xAB_CD_EF).to_le_bytes());
+        test_roundtrip(
+            i24::from_bits(0xAB_CD_EF),
+            &i24::from_bits(0xAB_CD_EF).to_le_bytes(),
+        );
+
         test_roundtrip(u24::new(0x12_34_56), &u24::new(0x12_34_56).to_le_bytes());
+        test_roundtrip(
+            i24::from_bits(0x12_34_56),
+            &i24::from_bits(0x12_34_56).to_le_bytes(),
+        );
 
         // Pick a byte boundary (80; test one below and one above to ensure we get the right number
         // of bytes)
@@ -3981,17 +4006,62 @@ mod borsh_tests {
             &[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x7F],
         );
         test_roundtrip(
+            i79::new(-1),
+            &[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x7F],
+        );
+
+        test_roundtrip(
             u80::MAX,
             &[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF],
         );
+        test_roundtrip(
+            i80::new(-1),
+            &[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF],
+        );
+
         test_roundtrip(
             u81::MAX,
             &[
                 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x01,
             ],
         );
+        test_roundtrip(
+            i81::new(-1),
+            &[
+                0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x01,
+            ],
+        );
 
-        // Test actual u128 and arbitrary u128 (which is a legal one, though not a predefined)
+        // Test MIN/MAX for signed integers
+        test_roundtrip(
+            i79::MAX,
+            &[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x3F],
+        );
+        test_roundtrip(
+            i80::MAX,
+            &[0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x7F],
+        );
+        test_roundtrip(
+            i81::MAX,
+            &[
+                0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0,
+            ],
+        );
+
+        test_roundtrip(
+            i79::MIN,
+            &[0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x40],
+        );
+        test_roundtrip(
+            i80::MIN,
+            &[0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x80],
+        );
+        test_roundtrip(
+            i81::MIN,
+            &[0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 1],
+        );
+
+        // Test actual u128 and arbitrary u128 (which is a legal one, though not predefined)
         test_roundtrip(
             u128::MAX,
             &[
@@ -4001,6 +4071,22 @@ mod borsh_tests {
         );
         test_roundtrip(
             UInt::<u128, 128>::MAX,
+            &[
+                0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
+                0xFF, 0xFF,
+            ],
+        );
+
+        // Test actual i128 and arbitrary i128 (which is a legal one, though not predefined)
+        test_roundtrip(
+            i128::new(-1),
+            &[
+                0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
+                0xFF, 0xFF,
+            ],
+        );
+        test_roundtrip(
+            Int::<i128, 128>::new(-1),
             &[
                 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
                 0xFF, 0xFF,
@@ -4021,20 +4107,31 @@ mod borsh_tests {
     #[test]
     fn test_schema_byte_count() {
         verify_byte_count_in_schema::<u1>(1, "u1");
+        verify_byte_count_in_schema::<i1>(1, "i1");
 
         verify_byte_count_in_schema::<u7>(1, "u7");
+        verify_byte_count_in_schema::<i7>(1, "i7");
 
         verify_byte_count_in_schema::<UInt<u8, 8>>(1, "u8");
+        verify_byte_count_in_schema::<Int<i8, 8>>(1, "i8");
+
         verify_byte_count_in_schema::<UInt<u32, 8>>(1, "u8");
+        verify_byte_count_in_schema::<Int<i32, 8>>(1, "i8");
 
         verify_byte_count_in_schema::<u9>(2, "u9");
+        verify_byte_count_in_schema::<i9>(2, "i9");
 
         verify_byte_count_in_schema::<u15>(2, "u15");
+        verify_byte_count_in_schema::<i15>(2, "i15");
+
         verify_byte_count_in_schema::<UInt<u128, 15>>(2, "u15");
+        verify_byte_count_in_schema::<Int<i128, 15>>(2, "i15");
 
         verify_byte_count_in_schema::<u63>(8, "u63");
+        verify_byte_count_in_schema::<i63>(8, "i63");
 
         verify_byte_count_in_schema::<u65>(9, "u65");
+        verify_byte_count_in_schema::<i65>(9, "i65");
     }
 }